// src/api/generated/models/ActivityReorderUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Blackwell Time API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ActivityReorderUpdate
 */
export interface ActivityReorderUpdate {
    /**
     * 
     * @type {string}
     * @memberof ActivityReorderUpdate
     */
    activityId?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityReorderUpdate
     */
    groupLabel?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ActivityReorderUpdate
     */
    sortOrder?: number;
}

/**
 * Check if a given object implements the ActivityReorderUpdate interface.
 */
export function instanceOfActivityReorderUpdate(value: object): value is ActivityReorderUpdate {
    return true;
}

export function ActivityReorderUpdateFromJSON(json: any): ActivityReorderUpdate {
    return ActivityReorderUpdateFromJSONTyped(json, false);
}

export function ActivityReorderUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ActivityReorderUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'activityId': json['activityId'] == null ? undefined : json['activityId'],
        'groupLabel': json['groupLabel'] == null ? undefined : json['groupLabel'],
        'sortOrder': json['sortOrder'] == null ? undefined : json['sortOrder'],
    };
}

export function ActivityReorderUpdateToJSON(json: any): ActivityReorderUpdate {
    return ActivityReorderUpdateToJSONTyped(json, false);
}

export function ActivityReorderUpdateToJSONTyped(value?: ActivityReorderUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'activityId': value['activityId'],
        'groupLabel': value['groupLabel'],
        'sortOrder': value['sortOrder'],
    };
}